requirements.txt
opencv_python_headless==4.5.4.60
PyExecJS==1.5.1
requests==2.27.1
